+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/hooks/useTimeScale.ts
