+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/panel.module.scss
